requirements.txt
aws-cdk-lib>=2.241.0,<3.0.0
constructs>=10.5.0,<11.0.0
aws-cdk.lambda-layer-kubectl-v32>=2.0.0,<3.0.0
